+++ tests/subscriber.cpp
@@ -10,10 +10,9 @@
 
 #include <servus/servus.h>
 
-#include <tests/newEvent_generated.h>
-#include <tests/newEvent_zeroeq_generated.h>
-
+#ifdef ZEROEQ_USE_FLATBUFFERS
 using namespace zeroeq::vocabulary;
+#endif
 
 BOOST_AUTO_TEST_CASE(construction)
 {
@@ -66,26 +65,27 @@ BOOST_AUTO_TEST_CASE(invalid_construction)
 BOOST_AUTO_TEST_CASE(subscribe)
 {
     zeroeq::Subscriber subscriber;
-    zeroeq::FBEvent echoEvent( ::zeroeq::vocabulary::EVENT_ECHO,
-                          std::bind( &test::onEchoEvent, std::placeholders::_1 ));
-    BOOST_CHECK( subscriber.subscribe( echoEvent ));
+    ::test::SerializablePtr echoEvent = ::test::getFBEchoInEvent(
+                    std::bind( &test::onEchoEvent, std::placeholders::_1 ));
+    BOOST_CHECK( subscriber.subscribe( *echoEvent ));
 }
 
 BOOST_AUTO_TEST_CASE(unsubscribe)
 {
     zeroeq::Subscriber subscriber;
-    zeroeq::FBEvent echoEvent( ::zeroeq::vocabulary::EVENT_ECHO,
-                          std::bind( &test::onEchoEvent, std::placeholders::_1 ));
-    BOOST_CHECK( subscriber.subscribe( echoEvent ));
-    BOOST_CHECK( subscriber.unsubscribe( echoEvent ));
+    ::test::SerializablePtr echoEvent = ::test::getFBEchoInEvent(
+                    std::bind( &test::onEchoEvent, std::placeholders::_1 ));
+    BOOST_CHECK( subscriber.subscribe( *echoEvent ));
+    BOOST_CHECK( subscriber.unsubscribe( *echoEvent ));
 }
 
 BOOST_AUTO_TEST_CASE(invalid_subscribe)
 {
     zeroeq::Subscriber subscriber;
-    test::Echo echo;
-    BOOST_CHECK( subscriber.subscribe( echo ));
-    BOOST_CHECK( !subscriber.subscribe( echo ));
+    ::test::SerializablePtr echo =
+            ::test::getFBEchoInEvent( ::zeroeq::EventFunc( ));
+    BOOST_CHECK( subscriber.subscribe( *echo ));
+    BOOST_CHECK( !subscriber.subscribe( *echo ));
 }
 
 BOOST_AUTO_TEST_CASE(test_invalid_unsubscribe)
@@ -102,13 +102,14 @@ BOOST_AUTO_TEST_CASE(test_invalid_unsubscribe)
 BOOST_AUTO_TEST_CASE(test_invalid_unsubscribe_different_event_objects)
 {
     zeroeq::Subscriber subscriber;
-    zeroeq::FBEvent echoEvent( ::zeroeq::vocabulary::EVENT_ECHO,
-                                ::zeroeq::EventFunc( ));
-    zeroeq::FBEvent newEvent( ::zeroeqtest::EVENT_NEWEVENT,
-                              ::zeroeq::EventFunc( ));
+    ::test::SerializablePtr echoEvent =
+            ::test::getFBEchoInEvent( ::zeroeq::EventFunc( ));
 
-    BOOST_CHECK( subscriber.subscribe( echoEvent ));
-    BOOST_CHECK( !subscriber.unsubscribe( newEvent ));
+    ::test::SerializablePtr emptyEvent =
+            ::test::getFBEmptyInEvent( ::zeroeq::EventFunc( ));
+
+    BOOST_CHECK( subscriber.subscribe( *echoEvent ));
+    BOOST_CHECK( !subscriber.unsubscribe( *emptyEvent ));
 }