+++ zeroeq/detail/fbevent.h
@@ -38,18 +38,14 @@ public:
     const void* getData() const
     {
         if( data )
-            return data.get();
+            return data;
         return parser.builder_.GetBufferPointer();
     }
 
     void setData( const uint8_t* data_, const size_t size_ )
     {
         parser.builder_.Clear();
-        // Since we are working on the same copy with new servus::Serializable
-        // based Flat Buffer events ( not memory per event ),
-        // we are copying the memory
-        data.reset( new uint8_t[ size_ ] );
-        memcpy( data.get(), data_, size_ );
+        data = data_;
         size = size_;
     }
 
@@ -59,8 +55,7 @@ public:
     flatbuffers::Parser parser;
 
     /** setData() uses this instead of fbb during deserialization */
-    std::shared_ptr< uint8_t > data;
-
+    const uint8_t* data;
     size_t size;
     EventFunc func;